refactor(jornades-pad): [refactor-dirs] Moved componets, libs and interfaces to src
