+++ src/app/[slug]/page.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
-import api from '@/app/libs/api.js';
-import { IMeta, IEvent } from '@/app/interfaces';
-import EventsList from '@/app/componnents/EventsList';
+import api from '@/libs/api.js';
+import { IMeta, IEvent } from '@/interfaces';
+import EventsList from '@/componnents/EventsList';
 
 import styles from '../page.module.scss';
