+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image'
 import { Inter } from 'next/font/google'
-import api from '@/app/libs/api.js';
+import api from '@/libs/api.js';
 import styles from './page.module.scss'
 
 const inter = Inter({ subsets: ['latin'] })
